The ASCII log writer can now write gzipped logs.
The setting can be set globally or per-filter with the gzip_level
value.  It is set to 0 by default, but if it is set to any value
1-9 it will cause Bro to write out logs gzipped at that gzip
compression level.  It also causes the writer to name the files
with ".gz" at the end.

This branch is still missing a test.